+++ src/adapter/bun/index.ts
@@ -192,6 +192,19 @@ const mergeRoutes = (r1: Routes, r2?: Routes) => {
 	return r1
 }
 
+// https://github.com/elysiajs/elysia/issues/1752
+// Bun.routes need to be sort for some reason?
+export const removeTrailingPath = (routes: Routes) => {
+	for (const key of Object.keys(routes)) {
+		if (key.charCodeAt(key.length - 1) === 47) {
+			routes[key.slice(0, -1)] = routes[key]
+			delete routes[key]
+		}
+	}
+
+	return routes
+}
+
 export const BunAdapter: ElysiaAdapter = {
 	...WebStandardAdapter,
 	name: 'bun',
@@ -326,6 +339,17 @@ export const BunAdapter: ElysiaAdapter = {
 				return staticRoutes as any
 			}
 
+			const routes = removeTrailingPath(
+				mergeRoutes(
+					mergeRoutes(
+						createStaticRoute(app.router.response),
+						mapRoutes(app)
+					),
+					// @ts-ignore
+					app.config.serve?.routes
+				)
+			)
+
 			const serve =
 				typeof options === 'object'
 					? ({
@@ -334,15 +358,7 @@ export const BunAdapter: ElysiaAdapter = {
 							idleTimeout: 30,
 							...(app.config.serve || {}),
 							...(options || {}),
-							// @ts-ignore
-							routes: mergeRoutes(
-								mergeRoutes(
-									createStaticRoute(app.router.response),
-									mapRoutes(app)
-								),
-								// @ts-ignore
-								app.config.serve?.routes
-							),
+							routes,
 							websocket: {
 								...(app.config.websocket || {}),
 								...(websocket || {}),
@@ -355,15 +371,7 @@ export const BunAdapter: ElysiaAdapter = {
 							reusePort: true,
 							idleTimeout: 30,
 							...(app.config.serve || {}),
-							// @ts-ignore
-							routes: mergeRoutes(
-								mergeRoutes(
-									createStaticRoute(app.router.response),
-									mapRoutes(app)
-								),
-								// @ts-ignore
-								app.config.serve?.routes
-							),
+							routes,
 							websocket: {
 								...(app.config.websocket || {}),
 								...(websocket || {})
@@ -395,11 +403,8 @@ export const BunAdapter: ElysiaAdapter = {
 			app.promisedModules.then(async () => {
 				if (typeof app.config.aot) app.compile()
 
-				app.server?.reload({
-					...serve,
-					fetch: app.fetch,
-					// @ts-ignore
-					routes: mergeRoutes(
+				const routes = removeTrailingPath(
+					mergeRoutes(
 						mergeRoutes(
 							await createStaticRoute(app.router.response, {
 								withAsync: true
@@ -409,6 +414,13 @@ export const BunAdapter: ElysiaAdapter = {
 						// @ts-ignore
 						app.config.serve?.routes
 					)
+				)
+
+				app.server?.reload({
+					...serve,
+					fetch: app.fetch,
+					// @ts-ignore
+					routes
 				})
 
 				Bun?.gc(false)